Move validation class
