+++ app/src/main/java/com/dubproductions/bracket/presentation/ui/screen/main/hosting/TournamentCreation.kt
@@ -26,8 +26,8 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.unit.dp
 import com.dubproductions.bracket.R
-import com.dubproductions.bracket.Type
-import com.dubproductions.bracket.Validation
+import com.dubproductions.bracket.utils.Type
+import com.dubproductions.bracket.utils.Validation
 import com.dubproductions.bracket.presentation.ui.components.dialogs.ReusableDialog
 import com.dubproductions.bracket.presentation.ui.state.TournamentCreationUIState
