+++ app/src/main/java/com/dubproductions/bracket/utils/Validation.kt
@@ -1,4 +1,4 @@
-package com.dubproductions.bracket
+package com.dubproductions.bracket.utils
 
 import android.util.Patterns
